image_to_music/requirements.txt
altair==5.4.1
astropy==6.1.4
astropy-iers-data==0.2024.9.30.0.32.59
attrs==24.2.0
audioread==3.0.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
gitdb==4.0.11
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
librosa==0.9.1
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
MIDIUtil==1.2.1
narwhals==1.9.1
numba==0.60.0
numpy==2.0.2
opencv-python-headless==4.10.0.84
pandas==2.2.3
pedalboard==0.5.9
Pillow==9.2.0
pooch==1.8.2
protobuf==3.20.3
pyarrow==17.0.0
pycparser==2.22
pydeck==0.9.1
pyerfa==2.0.1.4
Pympler==1.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
resampy==0.4.3
rich==13.9.2
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.14.1
semver==3.0.2
smmap==5.0.1
soundfile==0.12.1
streamlit==1.13.0
threadpoolctl==3.5.0
toml==0.10.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
validators==0.34.0